Homework

+++ Sesion-21/Ejemplo-set/datatypes.h
@@ -1,3 +1,10 @@
+/*
+ * datatypes.h
+ *
+ *  Created on: 15 mar 2023
+ *      Author: jluis
+ */
+
 #ifndef DATATYPES_H_
 
 #include <stdlib.h>
@@ -7,7 +14,7 @@ typedef int BOOL;
 typedef char * STRING;
 
 TYPE int_create(int i);
-TYPE float_create(int f);
+TYPE float_create(float f);
 TYPE double_create(int d);
 TYPE char_create(char c);
 TYPE string_create(STRING s);
@@ -17,10 +24,11 @@ float float_val(TYPE *e);
 double double_val(TYPE *e);
 int int_val(TYPE *e);
 
-// Convertir del tipo de dato int* o void* a COMPAREFUNC Y PRINTFUNC de funciones a apuntador
-typedef int (*COMPAREFUNC)(TYPE, TYPE);
+typedef int (*COMPAREFUNC)(TYPE,TYPE);
 typedef void (*PRINTFUNC)(TYPE);
 
 #define DATATYPES_H_
+
+
 
 #endif /* DATATYPES_H_ */
